Ensure the target directory exists during check.
For the rsync operation to work, the target directory must exist, or the
rsync will fail. For pipelines making use of multiple target locations,
this can be painful to maintain.

Now, the directory is created (if possible) during check, before
attempting to list the builds in the directory, allowing any path in a
writable location to be specified without creating it first.