Fix some issues running on Concourse.
The check script had over-aggressive quoting, which led to errors
checking for new versions of resources.

The in and out scripts were passing -v to rsync, causing a lot of extra
noise and slowing down any build logs. We don't need to see every file,
so we now turn off verbosity.